refactor: добавить регистрацию и авторизацию пользователя

+++ controllers/card.js
@@ -1,35 +1,48 @@
 const Card = require('../models/card');
-const { errorMessage, CREATED } = require('../utils/errorMessage');
+const { errorMessage } = require('../utils/errorMessage');
+const { NOT_FOUND_ERROR } = require('../errors/notFoundError');
+const { FORBIDDEN_ERROR } = require('../errors/forbiddenError');
+const { CREATED } = require('../utils/successes');
 
 // возвращает все карточки
-module.exports.getCard = (req, res) => {
+module.exports.getCard = (req, res, next) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
-    .catch((err) => errorMessage(err, req, res));
+    .catch((err) => errorMessage(err, req, res, next));
 };
 
 // создаёт карточку
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const ownerId = req.user._id;
 
   Card.create({ name, link, owner: ownerId })
     .then((card) => res.status(CREATED).send({ data: card }))
-    .catch((err) => errorMessage(err, req, res));
+    .catch((err) => errorMessage(err, req, res, next));
 };
 
 // удаляет карточку по идентификатору
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail()
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(() => {
+      throw new NOT_FOUND_ERROR('Карточка с таким id не найдена');
+    })
+    .then((card) => {
+      if (!card.owner.equals(req.user._id)) {
+        throw new FORBIDDEN_ERROR('Эту карточку нельзя удалить');
+      }
+      return Card.findByIdAndRemove(req.params.cardId);
+    })
     .then((card) => res.send({ data: card }))
-    .catch((err) => errorMessage(err, req, res));
+    .catch((err) => errorMessage(err, req, res, next));
 };
 
 // поставить лайк карточке
 module.exports.likeCard = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .orFail()
+    .orFail(() => {
+      throw new NOT_FOUND_ERROR('Карточка с таким id не найдена');
+    })
     .then((card) => res.send({ data: card }))
     .catch((err) => errorMessage(err, req, res));
 };
@@ -37,7 +50,9 @@ module.exports.likeCard = (req, res) => {
 // убрать лайк с карточки
 module.exports.dislikeCard = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .orFail()
+    .orFail(() => {
+      throw new NOT_FOUND_ERROR('Карточка с таким id не найдена');
+    })
     .then((card) => res.send({ data: card }))
     .catch((err) => errorMessage(err, req, res));
 };
